+++ crates/prover/benches/quotients.rs
@@ -4,9 +4,9 @@ use criterion::{black_box, Criterion};
 use itertools::Itertools;
 use stwo_prover::core::backend::CPUBackend;
 use stwo_prover::core::circle::SECURE_FIELD_CIRCLE_GEN;
-use stwo_prover::core::commitment_scheme::quotients::{ColumnSampleBatch, QuotientOps};
 use stwo_prover::core::fields::m31::BaseField;
 use stwo_prover::core::fields::qm31::SecureField;
+use stwo_prover::core::pcs::quotients::{ColumnSampleBatch, QuotientOps};
 use stwo_prover::core::poly::circle::{CanonicCoset, CircleEvaluation};
 use stwo_prover::core::poly::BitReversedOrder;
