+++ crates/prover/src/core/backend/avx512/blake2s.rs
@@ -4,11 +4,11 @@ use itertools::Itertools;
 
 use super::blake2s_avx::{compress16, set1, transpose_msgs, untranspose_states};
 use super::{AVX512Backend, VECS_LOG_SIZE};
-use crate::commitment_scheme::blake2_hash::Blake2sHash;
-use crate::commitment_scheme::blake2_merkle::Blake2sMerkleHasher;
-use crate::commitment_scheme::ops::{MerkleHasher, MerkleOps};
 use crate::core::backend::{Col, Column, ColumnOps};
 use crate::core::fields::m31::BaseField;
+use crate::core::vcs::blake2_hash::Blake2sHash;
+use crate::core::vcs::blake2_merkle::Blake2sMerkleHasher;
+use crate::core::vcs::ops::{MerkleHasher, MerkleOps};
 
 impl ColumnOps<Blake2sHash> for AVX512Backend {
     type Column = Vec<Blake2sHash>;
