+++ crates/prover/src/core/prover/mod.rs
@@ -4,25 +4,25 @@ use tracing::{span, Level};
 
 use super::air::AirProver;
 use super::backend::Backend;
-use super::commitment_scheme::{CommitmentSchemeProof, TreeVec};
 use super::fri::FriVerificationError;
+use super::pcs::{CommitmentSchemeProof, TreeVec};
 use super::poly::circle::{CanonicCoset, SecureCirclePoly, MAX_CIRCLE_DOMAIN_LOG_SIZE};
 use super::proof_of_work::ProofOfWorkVerificationError;
 use super::ColumnVec;
-use crate::commitment_scheme::blake2_hash::Blake2sHasher;
-use crate::commitment_scheme::blake2_merkle::Blake2sMerkleHasher;
-use crate::commitment_scheme::hasher::Hasher;
-use crate::commitment_scheme::ops::MerkleOps;
-use crate::commitment_scheme::verifier::MerkleVerificationError;
 use crate::core::air::{Air, AirExt, AirProverExt};
 use crate::core::backend::CPUBackend;
 use crate::core::channel::{Blake2sChannel, Channel as ChannelTrait};
 use crate::core::circle::CirclePoint;
-use crate::core::commitment_scheme::{CommitmentSchemeProver, CommitmentSchemeVerifier};
 use crate::core::fields::m31::BaseField;
 use crate::core::fields::qm31::SecureField;
+use crate::core::pcs::{CommitmentSchemeProver, CommitmentSchemeVerifier};
 use crate::core::poly::circle::CircleEvaluation;
 use crate::core::poly::BitReversedOrder;
+use crate::core::vcs::blake2_hash::Blake2sHasher;
+use crate::core::vcs::blake2_merkle::Blake2sMerkleHasher;
+use crate::core::vcs::hasher::Hasher;
+use crate::core::vcs::ops::MerkleOps;
+use crate::core::vcs::verifier::MerkleVerificationError;
 use crate::core::ComponentVec;
 
 type Channel = Blake2sChannel;
