+++ crates/prover/src/core/vcs/ops.rs
