+++ crates/prover/src/core/vcs/utils.rs
