+++ crates/prover/src/hash_functions/poseidon.rs
@@ -2,9 +2,9 @@ use core::fmt;
 
 use num_traits::{One, Zero};
 
-use crate::commitment_scheme::hasher::{self, Hasher, Name};
 use crate::core::fields::m31::BaseField;
 use crate::core::fields::FieldExpOps;
+use crate::core::vcs::hasher::{self, Hasher, Name};
 use crate::math::matrix::{RowMajorMatrix, SquareMatrix};
 
 const POSEIDON_WIDTH: usize = 24; // in BaseField elements.
@@ -223,8 +223,8 @@ impl Hasher for PoseidonHasher {
 #[cfg(test)]
 mod tests {
     use super::{PoseidonHasher, POSEIDON_CAPACITY};
-    use crate::commitment_scheme::hasher::Hasher;
     use crate::core::fields::m31::BaseField;
+    use crate::core::vcs::hasher::Hasher;
     use crate::hash_functions::poseidon::PoseidonHash;
     use crate::m31;
